backend/agent-coordination-service/requirements.txt
# backend/agent-coordination-service/requirements.txt
fastapi
uvicorn
openai
pinecone-client
python-dotenv
